chore: extract Building Styles configuration
Extracted the `BUILDING_STYLES` configuration and `BuildingStyle` interface from `BuildingRenderer.ts` into a new central configuration file `claudeville/src/config/buildingStyles.ts`.

Unified the building styles across the codebase by:
- Consolidating the duplicate definition found in `claudeville/src/presentation/react/world/styles.ts`.
- Standardizing property names (e.g., using `roundRoof` instead of `hasRoundRoof`).
- Updating the Canvas-based `BuildingRenderer.ts` and React components to use the new central configuration.
- Updating type definitions to maintain type safety.

This improves maintainability by separating data from logic and eliminating duplication.
